✨ tweak: move message responder to message event file
