+++ src/events/message.js
@@ -1,4 +1,4 @@
-const { Client, Events } = require("discord.js");
+const { Client, Events, GatewayIntentBits, Partials } = require("discord.js");
 const { sendEmbedMessage, serverTime } = require("../utils/core");
 const { version } = require("../../package.json");
 
@@ -12,14 +12,17 @@ const client = new Client({
 	partials: [Partials.Channel, Partials.Message],
 });
 
+// discord bot env
+const {
+	DISCORD_SUPPORT_ROLE_ID,
+	ASKAI_CHANNEL
+} = process.env;
+const roleIDs = DISCORD_SUPPORT_ROLE_ID.split(",");
+
 module.exports = {
     name: Events.MessageCreate,
     once: false,
     execute(message) {
-
-      // get the details from user who send command
-      const member = message.member;
-      const mention = message.mentions;
 
       // prevent someone from sending DM to the bot.
       if (message.author.bot) return;
@@ -53,6 +56,22 @@ module.exports = {
         message.reply({
           embeds: [sendEmbedMessage(config.reminder_mention)],
         });
+      }
+
+      // get the details from user who send command
+      const member = message.member;
+      const mention = message.mentions;
+
+      // check if the message has mentioned a team member and not a reply
+      if (mention.users.first() && !message.reference) {
+        let mentioned = message.guild.members.cache.get(mention.users.first().id)
+        if (mentioned.roles.cache.hasAny(...roleIDs) && !member.roles.cache.hasAny(...roleIDs)) {
+          message.reply({
+            embeds: [sendEmbedMessage(`We have moved to a community driven discord support model.\n\nYou can ask me all things thirdweb in the <#${ASKAI_CHANNEL}> channel. Use the command \`!askai\` or \`!ask\` followed by your question to get started.`)],
+          }).then(msg => {
+            setTimeout(() => msg.delete(), 60000)
+          })
+        }
       }
 
     },
